[Seq] Merge Seq-to-SV lowering passes
The pattern rewriter can be applied alongside the FIR register lowering transforms to eliminate the need for two separate passes.
