+++ include/circt-c/Conversion.h
@@ -1,0 +1,12 @@
+//===-- circt-c/Dialect.h - C API for dialect conversions ---------*- C -*-===//
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef CIRCT_C_CONVERSION_H
+#define CIRCT_C_CONVERSION_H
+
+#include "mlir-c/Support.h"
+
+#include "mlir/Conversion/Passes.capi.h.inc"
+
+#endif // CIRCT_C_CONVERSION_H
