+++ lib/CAPI/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(Conversion)
 add_subdirectory(ExportFIRRTL)
 add_subdirectory(ExportVerilog)
 add_subdirectory(Dialect)
