+++ lib/CAPI/Conversion/CMakeLists.txt
@@ -1,0 +1,33 @@
+add_mlir_public_c_api_library(CIRCTCAPIConversion
+  Passes.cpp
+
+  LINK_LIBS PUBLIC
+  CIRCTAffineToLoopSchedule
+  CIRCTArcToLLVM
+  CIRCTCalyxToFSM
+  CIRCTCalyxToHW
+  CIRCTCombToArith
+  CIRCTCombToLLVM
+  CIRCTConvertToArcs
+  CIRCTDCToHW
+  CIRCTExportChiselInterface
+  CIRCTExportVerilog
+  CIRCTFIRRTLToHW
+  CIRCTFSMToSV
+  CIRCTHandshakeToDC
+  CIRCTHandshakeToHW
+  CIRCTHWArithToHW
+  CIRCTHWToLLHD
+  CIRCTHWToLLVM
+  CIRCTHWToSV
+  CIRCTHWToSystemC
+  CIRCTLLHDToLLVM
+  CIRCTLoopScheduleToCalyx
+  CIRCTMooreToCore
+  CIRCTPipelineToHW
+  CIRCTSCFToCalyx
+  CIRCTSeqToSV
+  CIRCTStandardToHandshake
+  CIRCTVerifToSV
+)
+
