+++ lib/Dialect/Seq/Transforms/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_circt_dialect_library(CIRCTSeqTransforms
   ExternalizeClockGate.cpp
   LowerSeqHLMem.cpp
-  LowerSeqToSV.cpp
   LowerSeqFIFO.cpp
 
   DEPENDS
