+++ lib/Firtool/Firtool.cpp
@@ -210,7 +210,7 @@ LogicalResult firtool::populateHWToSV(mlir::PassManager &pm,
                                                opt.etcDisableModuleInlining));
 
   pm.addPass(seq::createExternalizeClockGatePass(opt.clockGateOpts));
-  pm.addPass(circt::createSeqFIRRTLLowerToSVPass(
+  pm.addPass(circt::createLowerSeqToSVPass(
       {/*disableRandomization=*/!opt.isRandomEnabled(
            FirtoolOptions::RandomKind::Reg),
        /*emitSeparateAlwaysBlocks=*/
